requirements.txt
torch
torchvision
transformers
pillow
pandas
scikit-learn
